Extract `//cuttlefish/common/libs/utils:disk_usage`
This breaks part of a circular dependency between `files.cpp` and
`subprocess.cpp`.

Bug: b/407763557
