+++ base/cvd/cuttlefish/common/libs/utils/files.cpp
@@ -78,8 +78,6 @@
 
 namespace cuttlefish {
 
-static constexpr char kWhitespaceCharacters[] = " \n\t\r\v\f";
-
 bool FileExists(const std::string& path, bool follow_symlinks) {
   struct stat st {};
   return (follow_symlinks ? stat : lstat)(path.c_str(), &st) == 0;
@@ -617,43 +615,6 @@ FileSizes SparseFileSizes(const std::string& path) {
 bool FileIsSocket(const std::string& path) {
   struct stat st {};
   return stat(path.c_str(), &st) == 0 && S_ISSOCK(st.st_mode);
-}
-
-// return unit determined by the `--block-size` argument
-Result<std::size_t> GetDiskUsage(const std::string& path,
-                                 const std::string& size_arg) {
-  Command du_cmd("du");
-  du_cmd.AddParameter("-s");  // summarize, only output total
-  du_cmd.AddParameter(
-      "--apparent-size");  // apparent size rather than device usage
-  du_cmd.AddParameter("--block-size=" + size_arg);
-  du_cmd.AddParameter(path);
-
-  std::string out;
-  std::string err;
-  int return_code = RunWithManagedStdio(std::move(du_cmd), nullptr, &out, &err);
-  CF_EXPECTF(return_code == 0, "Failed to run `du` command.  stderr: {}", err);
-  CF_EXPECTF(!out.empty(), "No output read from `du` command. stderr: {}", err);
-  std::vector<std::string> split_out =
-      android::base::Tokenize(out, kWhitespaceCharacters);
-  CF_EXPECTF(!split_out.empty(),
-             "No valid output read from `du` command in \"{}\"", out);
-  std::string total = split_out.front();
-
-  std::size_t result;
-  CF_EXPECTF(android::base::ParseUint(total, &result),
-             "Failure parsing \"{}\" to integer.", total);
-  return result;
-}
-
-Result<std::size_t> GetDiskUsageBytes(const std::string& path) {
-  return CF_EXPECTF(GetDiskUsage(path, "1"),
-                    "Unable to determine disk usage of file \"{}\"", path);
-}
-
-Result<std::size_t> GetDiskUsageGigabytes(const std::string& path) {
-  return CF_EXPECTF(GetDiskUsage(path, "1G"),
-                    "Unable to determine disk usage of file \"{}\"", path);
 }
 
 /**
