+++ base/cvd/cuttlefish/host/commands/cvd/cache/BUILD.bazel
@@ -11,8 +11,9 @@ cc_library(
     hdrs = ["cache.h"],
     copts = COPTS + [ "-Werror=sign-compare" ],
     deps = [
-        "//cuttlefish/common/libs/utils:result",
         "//cuttlefish/common/libs/utils",
+        "//cuttlefish/common/libs/utils:disk_usage",
+        "//cuttlefish/common/libs/utils:result",
         "//libbase",
         "@fmt",
         "@jsoncpp",
