+++ base/cvd/cuttlefish/host/commands/cvd/cache/cache.cpp
@@ -27,6 +27,7 @@
 #include <android-base/strings.h>
 #include <fmt/format.h>
 
+#include "cuttlefish/common/libs/utils/disk_usage.h"
 #include "cuttlefish/common/libs/utils/files.h"
 #include "cuttlefish/common/libs/utils/result.h"
